moved Register and Login files back under Components
